// src/tsb/ast/nodes.generated.ts
export interface ArgumentsNode extends NodeBase {
  kind: "arguments";
  children: (ArrayLiteralNode | ArrowFunctionNode | AssignmentExpressionNode | BinaryExpressionNode | BooleanNode | CallExpressionNode | IdentifierNode | MatchExpressionNode | MemberExpressionNode | NullLiteralNode | NumberNode | ObjectLiteralNode | ParenthesisedExpressionNode | StringNode | SubscriptExpressionNode | TemplateStringNode | TernaryExpressionNode | UnaryExpressionNode | UndefinedLiteralNode)[];
}

/** Grammar node `array_literal`. */
export interface ArrayLiteralNode extends NodeBase {
  kind: "array_literal";
  children: (ArrayLiteralNode | ArrowFunctionNode | AssignmentExpressionNode | BinaryExpressionNode | BooleanNode | CallExpressionNode | IdentifierNode | MatchExpressionNode | MemberExpressionNode | NullLiteralNode | NumberNode | ObjectLiteralNode | ParenthesisedExpressionNode | StringNode | SubscriptExpressionNode | TemplateStringNode | TernaryExpressionNode | UnaryExpressionNode | UndefinedLiteralNode)[];
}

/** Grammar node `array_type`. */
export interface ArrayTypeNode extends NodeBase {
  kind: "array_type";
  children: ArrayTypeNode | BooleanNode | FunctionTypeNode | GenericTypeNode | NamedTypeNode | NullLiteralNode | NumberNode | ObjectTypeNode | PrimitiveTypeNode | SelfTypeNode | StringNode | TupleTypeNode | UndefinedLiteralNode;
}

/** Grammar node `arrow_function`. */
export interface ArrowFunctionNode extends NodeBase {
  kind: "arrow_function";
  body: ArrayLiteralNode | ArrowFunctionNode | AssignmentExpressionNode | BinaryExpressionNode | BooleanNode | CallExpressionNode | IdentifierNode | MatchExpressionNode | MemberExpressionNode | NullLiteralNode | NumberNode | ObjectLiteralNode | ParenthesisedExpressionNode | StatementBlockNode | StringNode | SubscriptExpressionNode | TemplateStringNode | TernaryExpressionNode | UnaryExpressionNode | UndefinedLiteralNode;
  parameters: ArrowParametersNode | IdentifierNode;
}

/** Grammar node `arrow_parameters`. */
export interface ArrowParametersNode extends NodeBase {
  kind: "arrow_parameters";
  children: (FormalParameterNode)[];
}

/** Grammar node `assignment_expression`. */
export interface AssignmentExpressionNode extends NodeBase {
  kind: "assignment_expression";
  left: IdentifierNode | MemberExpressionNode | SubscriptExpressionNode;
  right: ArrayLiteralNode | ArrowFunctionNode | AssignmentExpressionNode | BinaryExpressionNode | BooleanNode | CallExpressionNode | IdentifierNode | MatchExpressionNode | MemberExpressionNode | NullLiteralNode | NumberNode | ObjectLiteralNode | ParenthesisedExpressionNode | StringNode | SubscriptExpressionNode | TemplateStringNode | TernaryExpressionNode | UnaryExpressionNode | UndefinedLiteralNode;
}

/** Grammar node `attribute`. */
export interface AttributeNode extends NodeBase {
  kind: "attribute";
  children: (AttributeItemNode)[];
}

/** Grammar node `attribute_arguments`. */
export interface AttributeArgumentsNode extends NodeBase {
  kind: "attribute_arguments";
  children: (AttributeCallNode | BooleanNode | IdentifierNode | NumberNode | StringNode)[];
}

/** Grammar node `attribute_call`. */
export interface AttributeCallNode extends NodeBase {
  kind: "attribute_call";
  name: IdentifierNode;
  children: (AttributeCallNode | BooleanNode | IdentifierNode | NumberNode | StringNode)[];
}

/** Grammar node `attribute_item`. */
export interface AttributeItemNode extends NodeBase {
  kind: "attribute_item";
  arguments?: AttributeArgumentsNode | undefined;
  name: IdentifierNode;
}

/** Grammar node `attributed_declaration`. */
export interface AttributedDeclarationNode extends NodeBase {
  kind: "attributed_declaration";
  children: (AttributeNode | FunctionDeclarationNode | ImplDeclarationNode | StructDeclarationNode | TraitDeclarationNode)[];
}

/** Grammar node `binary_expression`. */
export interface BinaryExpressionNode extends NodeBase {
  kind: "binary_expression";
  children: (ArrayLiteralNode | ArrowFunctionNode | BinaryExpressionNode | BooleanNode | CallExpressionNode | IdentifierNode | MatchExpressionNode | MemberExpressionNode | NullLiteralNode | NumberNode | ObjectLiteralNode | ParenthesisedExpressionNode | StringNode | SubscriptExpressionNode | TemplateStringNode | TernaryExpressionNode | UnaryExpressionNode | UndefinedLiteralNode)[];
}

/** Grammar node `binding_pattern`. */
export interface BindingPatternNode extends NodeBase {
  kind: "binding_pattern";
}

/** Grammar node `boolean`. */
export interface BooleanNode extends NodeBase {
  kind: "boolean";
}

/** Grammar node `call_expression`. */
export interface CallExpressionNode extends NodeBase {
  kind: "call_expression";
  arguments: ArgumentsNode;
  function: ArrayLiteralNode | ArrowFunctionNode | BinaryExpressionNode | BooleanNode | CallExpressionNode | IdentifierNode | MatchExpressionNode | MemberExpressionNode | NullLiteralNode | NumberNode | ObjectLiteralNode | ParenthesisedExpressionNode | StringNode | SubscriptExpressionNode | TemplateStringNode | TernaryExpressionNode | UnaryExpressionNode | UndefinedLiteralNode;
}

/** Grammar node `export_statement`. */
export interface ExportStatementNode extends NodeBase {
  kind: "export_statement";
  source?: StringNode | undefined;
  children: IdentifierNode | NamedImportsNode | undefined;
}

/** Grammar node `formal_parameter`. */
export interface FormalParameterNode extends NodeBase {
  kind: "formal_parameter";
  default?: ArrayLiteralNode | ArrowFunctionNode | AssignmentExpressionNode | BinaryExpressionNode | BooleanNode | CallExpressionNode | IdentifierNode | MatchExpressionNode | MemberExpressionNode | NullLiteralNode | NumberNode | ObjectLiteralNode | ParenthesisedExpressionNode | StringNode | SubscriptExpressionNode | TemplateStringNode | TernaryExpressionNode | UnaryExpressionNode | UndefinedLiteralNode | undefined;
  name: IdentifierNode;
  type?: ArrayTypeNode | BooleanNode | FunctionTypeNode | GenericTypeNode | NamedTypeNode | NullLiteralNode | NumberNode | ObjectTypeNode | PrimitiveTypeNode | SelfTypeNode | StringNode | TupleTypeNode | UndefinedLiteralNode | UnionTypeNode | undefined;
}

/** Grammar node `function_declaration`. */
export interface FunctionDeclarationNode extends NodeBase {
  kind: "function_declaration";
  body: StatementBlockNode;
  generics?: TypeParametersNode | undefined;
  name: IdentifierNode;
  parameters: ArrowParametersNode;
  return_type?: ArrayTypeNode | BooleanNode | FunctionTypeNode | GenericTypeNode | NamedTypeNode | NullLiteralNode | NumberNode | ObjectTypeNode | PrimitiveTypeNode | SelfTypeNode | StringNode | TupleTypeNode | UndefinedLiteralNode | UnionTypeNode | undefined;
}

/** Grammar node `function_type`. */
export interface FunctionTypeNode extends NodeBase {
  kind: "function_type";
  children: (ArrayTypeNode | BooleanNode | FunctionTypeNode | GenericTypeNode | NamedTypeNode | NullLiteralNode | NumberNode | ObjectTypeNode | PrimitiveTypeNode | SelfTypeNode | StringNode | TupleTypeNode | TypeParameterDeclNode | UndefinedLiteralNode | UnionTypeNode)[];
}

/** Grammar node `generic_type`. */
export interface GenericTypeNode extends NodeBase {
  kind: "generic_type";
  arguments: TypeArgumentsNode;
  name: TypeIdentifierNode;
}
// ...
export interface ImplBodyNode extends NodeBase {
  kind: "impl_body";
  children: (ImplMethodNode)[];
}

/** Grammar node `impl_declaration`. */
export interface ImplDeclarationNode extends NodeBase {
  kind: "impl_declaration";
  body: ImplBodyNode;
  first: TypeIdentifierNode;
  target?: TypeIdentifierNode | undefined;
  trait_args?: TypeArgumentsNode | undefined;
}

/** Grammar node `impl_method`. */
export interface ImplMethodNode extends NodeBase {
  kind: "impl_method";
  body: StatementBlockNode;
  generics?: TypeParametersNode | undefined;
  name: IdentifierNode;
  parameters: ArrowParametersNode;
  return_type?: ArrayTypeNode | BooleanNode | FunctionTypeNode | GenericTypeNode | NamedTypeNode | NullLiteralNode | NumberNode | ObjectTypeNode | PrimitiveTypeNode | SelfTypeNode | StringNode | TupleTypeNode | UndefinedLiteralNode | UnionTypeNode | undefined;
  children: (AttributeNode)[];
}

/** Grammar node `import_clause`. */
export interface ImportClauseNode extends NodeBase {
  kind: "import_clause";
  default?: IdentifierNode | undefined;
  children: NamedImportsNode | NamespaceImportNode | undefined;
}

/** Grammar node `import_specifier`. */
export interface ImportSpecifierNode extends NodeBase {
  kind: "import_specifier";
  alias?: IdentifierNode | undefined;
  name: IdentifierNode;
}

/** Grammar node `import_statement`. */
export interface ImportStatementNode extends NodeBase {
  kind: "import_statement";
  source: StringNode;
  children: ImportClauseNode;
}

/** Grammar node `literal_pattern`. */
export interface LiteralPatternNode extends NodeBase {
  kind: "literal_pattern";
  children: BooleanNode | NullLiteralNode | NumberNode | StringNode | UndefinedLiteralNode;
}

/** Grammar node `match_arm`. */
export interface MatchArmNode extends NodeBase {
  kind: "match_arm";
  body: ArrayLiteralNode | ArrowFunctionNode | AssignmentExpressionNode | BinaryExpressionNode | BooleanNode | CallExpressionNode | IdentifierNode | MatchExpressionNode | MemberExpressionNode | NullLiteralNode | NumberNode | ObjectLiteralNode | ParenthesisedExpressionNode | StringNode | SubscriptExpressionNode | TemplateStringNode | TernaryExpressionNode | UnaryExpressionNode | UndefinedLiteralNode;
  pattern: BindingPatternNode | LiteralPatternNode | ObjectPatternNode | StructPatternNode | WildcardPatternNode;
}
// ...
export interface NamedImportsNode extends NodeBase {
  kind: "named_imports";
  children: (ImportSpecifierNode)[];
}

/** Grammar node `named_type`. */
export interface NamedTypeNode extends NodeBase {
  kind: "named_type";
  children: TypeIdentifierNode;
}

/** Grammar node `namespace_import`. */
export interface NamespaceImportNode extends NodeBase {
  kind: "namespace_import";
  children: IdentifierNode;
}
// ...
export interface StructDeclarationNode extends NodeBase {
  kind: "struct_declaration";
  body: StructBodyNode;
  generics?: TypeParametersNode | undefined;
  name: TypeIdentifierNode;
}

/** Grammar node `struct_field`. */
export interface StructFieldNode extends NodeBase {
  kind: "struct_field";
  name: IdentifierNode;
  type: ArrayTypeNode | BooleanNode | FunctionTypeNode | GenericTypeNode | NamedTypeNode | NullLiteralNode | NumberNode | ObjectTypeNode | PrimitiveTypeNode | SelfTypeNode | StringNode | TupleTypeNode | UndefinedLiteralNode | UnionTypeNode;
  children: (AttributeNode)[];
}
// ...
export interface TraitDeclarationNode extends NodeBase {
  kind: "trait_declaration";
  body: TraitBodyNode;
  generics?: TypeParametersNode | undefined;
  name: TypeIdentifierNode;
}

/** Grammar node `trait_method`. */
export interface TraitMethodNode extends NodeBase {
  kind: "trait_method";
  body?: StatementBlockNode | undefined;
  generics?: TypeParametersNode | undefined;
  name: IdentifierNode;
  parameters: ArrowParametersNode;
  return_type?: ArrayTypeNode | BooleanNode | FunctionTypeNode | GenericTypeNode | NamedTypeNode | NullLiteralNode | NumberNode | ObjectTypeNode | PrimitiveTypeNode | SelfTypeNode | StringNode | TupleTypeNode | UndefinedLiteralNode | UnionTypeNode | undefined;
}
// ...
export type AstNode =
  | ArgumentsNode
  | ArrayLiteralNode
  | ArrayTypeNode
  | ArrowFunctionNode
  | ArrowParametersNode
  | AssignmentExpressionNode
  | AttributeNode
  | AttributeArgumentsNode
  | AttributeCallNode
  | AttributeItemNode
  | AttributedDeclarationNode
  | BinaryExpressionNode
  | BindingPatternNode
  | BooleanNode
  | CallExpressionNode
  | ExportStatementNode
  | FormalParameterNode
  | FunctionDeclarationNode
  | FunctionTypeNode
  | GenericTypeNode
  | IfStatementNode
  | ImplBodyNode
  | ImplDeclarationNode
  | ImplMethodNode
  | ImportClauseNode
  | ImportSpecifierNode
  | ImportStatementNode
  | LiteralPatternNode
  | MatchArmNode
  | MatchExpressionNode
  | MemberExpressionNode
  | NamedImportsNode
  | NamedTypeNode
  | NamespaceImportNode
  | ObjectLiteralNode
  | ObjectPatternNode
  | ObjectPropertyNode
  | ObjectTypeNode
  | ParenthesisedExpressionNode
  | PatternBindNode
  | PatternBodyNode
  | PatternCheckNode
  | PatternShorthandNode
  | PrimitiveTypeNode
  | PropertySignatureNode
  | ReturnStatementNode
  | ShorthandPropertyNode
  | SourceFileNode
  | SpreadElementNode
  | StatementBlockNode
  | StringNode
  | StructBodyNode
  | StructDeclarationNode
  | StructFieldNode
  | StructPatternNode
  | SubscriptExpressionNode
  | TemplateStringNode
  | TemplateSubstitutionNode
  | TernaryExpressionNode
  | TraitBodyNode
  | TraitDeclarationNode
  | TraitMethodNode
  | TupleTypeNode
  | TypeAliasNode
  | TypeArgumentsNode
  | TypeParameterNode
  | TypeParameterDeclNode
  | TypeParametersNode
  | UnaryExpressionNode
  | UnionTypeNode
  | VariableDeclarationNode
  | BlockCommentNode
  | EscapeSequenceNode
  | IdentifierNode
  | LineCommentNode
  | NullLiteralNode
  | NumberNode
  | SelfTypeNode
  | StringFragmentDoubleNode
  | StringFragmentSingleNode
  | TemplateCharsNode
  | TypeIdentifierNode
  | UndefinedLiteralNode
  | WildcardPatternNode